--- alembic/versions/0c5cf09f1fc4_initial_revision.py
@@ -1,21 +1,0 @@
-"""initial revision
-
-Revision ID: 0c5cf09f1fc4
-Revises:
-Create Date: 2018-05-28 15:01:48.308789
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '0c5cf09f1fc4'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
