+++ alembic/versions/88eb5854154f_add_id_group_in_request_sets.py
@@ -1,0 +1,24 @@
+"""add id_group in request_sets
+
+Revision ID: 88eb5854154f
+Revises: 
+Create Date: 2020-07-07 11:02:39.321300
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '88eb5854154f'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    op.add_column('request_sets', sa.Column('id_group', sa.TEXT))
+
+
+def downgrade():
+    op.op.drop_column('request_sets', 'id_group')
